Redefine the interface of IExpiry.Reminings and test refactorings
